requirements.txt
streamlit
groq
python-dotenv
pandas
google-play-scraper
langdetect
emoji
markdown
